Fixed a recently introduced mismatch between format specification and argument list in the ip-mib implementation (backported r19779 from the trunk).
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/branches/V5-5-patches@19791 06827809-a52a-0410-b366-d66718629ded

+++ agent/mibgroup/ip-mib/data_access/arp_linux.c
@@ -336,7 +336,7 @@ fillup_entry_info(netsnmp_arp_entry *entry, struct nlmsghdr *nlmp)
         length = nlmp->nlmsg_len - NLMSG_LENGTH(sizeof(*rtmp));
         if (length < 0) {
             snmp_log(LOG_ERR, "netlink message length %d < %d is invalid\n",
-                     nlmp->nlmsg_len, NLMSG_LENGTH(sizeof(*rtmp)));
+                     nlmp->nlmsg_len, (int)NLMSG_LENGTH(sizeof(*rtmp)));
             return -1;
         }
         /*
